Fix an unhandled exception
When the first chunk is very short and doesn't contain (0002,0010), the metainfo_listener is registered a second time.
When a binary element such as pixel data is processed and emits a event that contains only raw data, the listener
attempts to access event.element.tag, but element is undefined.

This prevents the second registration, but also guards against using a missing element property of the event.